% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Death_Row.R
\docType{package}
\name{Death_Row}
\alias{Death_Row}
\alias{Death_Row-package}
\title{Prisoners for Death Row Inmates}
\format{
A dataframe with 7335 elements
\describe{
\item{status}{If the person is still on death row}
\item{state}{state imprisoned in}
\item{birthyear}{Year the prisoner was born}
\item{age}{Age of the prisoner when they were sentenced}
\item{sentencing_year}{Year Sentenced}
\item{race}{Race of prisoner}
\item{gender}{Gender of prisoner}
\item{currently_on_death_row}{Whether or not prisoner is on death row}
\item{executed}{Whether or not prisoner has been executed}
\item{current_sentence}{Types of Sentences}
}
}
\source{
\url{https://tinyletter.com/data-is-plural/letters/data-is-plural-2019-12-11-edition}
}
\usage{
Death_Row
}
\description{
This package contains prisoner information regarding
inmates on death row

This package contains prisoner information regarding
inmates on death row
as of October, 2020
}
\keyword{datasets}
